# Performance improvement in serializer.