@blowin 2019-08-08
This is may be useful for cases. when often create new small arrays. There will be less memory allocated, which will speed up garbage collection.

It is generally difficult to measure real performance problems in garbage collection. Since they occur while the application is running, they can be critical in certain situations.